refactor: merge CLI and RPC
