+++ package.json
@@ -1,47 +1,136 @@
 {
-    "private": true,
-    "name": "exchange-json-rpc",
-    "description": "A JSON-RPC 2.0 specification compliant server for Exchanges.",
+    "name": "@arkecosystem/exchange-json-rpc-cli",
+    "version": "1.0.6",
+    "description": "A JSON-RPC 2.0 specification compliant server for Exchanges to interact with the ARK Blockchain.",
+    "license": "MIT",
+    "author": "Brian Faust <brian@ark.io>",
+    "files": [
+        "/bin",
+        "/dist",
+        "/oclif.manifest.json"
+    ],
+    "main": "dist/index",
+    "types": "dist/index",
+    "bin": {
+        "exchange-json-rpc": "./bin/run"
+    },
     "scripts": {
-        "lerna": "./node_modules/lerna/cli.js",
-        "setup": "yarn && yarn bootstrap && yarn build",
-        "setup:clean": "yarn && yarn clean && yarn bootstrap && yarn build",
-        "bootstrap": "yarn lerna bootstrap",
-        "clean": "yarn lerna clean --yes",
-        "build": "yarn lerna run build",
-        "lint": "./node_modules/tslint/bin/tslint -c ./tslint.json -p ./tslint-config.json './packages/**/*/src/**/*.ts' --fix",
+        "build": "yarn clean && yarn compile",
+        "build:watch": "yarn clean && yarn compile -w",
+        "clean": "del dist",
+        "compile": "./node_modules/typescript/bin/tsc",
+        "exchange-json-rpc": "./bin/run",
         "format": "yarn lint && yarn prettier",
-        "prettier": "prettier --write \"./*.{ts,js,json,md}\" \"./packages/**/*.{ts,js,json,md}\" \"./__tests__/**/*.{ts,js,json,md}\"",
-        "test": "yarn lerna run test",
-        "version": "cross-env-shell ./scripts/version.sh"
+        "lint": "./node_modules/tslint/bin/tslint -c ../../tslint.json -p ../../tslint-config.json './src/**/*.ts' --fix",
+        "prepack": "./node_modules/.bin/oclif-dev manifest && npm shrinkwrap",
+        "postpack": "rm -f oclif.manifest.json",
+        "prepublishOnly": "yarn build",
+        "prettier": "prettier --write \"./*.{ts,js,json,md}\" \"./src/**/*.{ts,js,json,md}\" \"./__tests__/**/*.{ts,js,json,md}\"",
+        "test": "jest"
+    },
+    "dependencies": {
+        "@arkecosystem/crypto": "^2.6.0-next.0",
+        "@arkecosystem/exchange-json-rpc": "^1.0.6",
+        "@arkecosystem/peers": "^0.2.1",
+        "@faustbrian/foreman": "^0.1.3",
+        "@hapi/boom": "^9.0.0",
+        "@hapi/hapi": "^18.3.1",
+        "@hapist/json-rpc": "^0.2.0",
+        "@hapist/whitelist": "^0.1.0",
+        "@oclif/command": "^1.5.14",
+        "@oclif/config": "^1.13.0",
+        "@oclif/plugin-autocomplete": "^0.1.0",
+        "@oclif/plugin-commands": "^1.2.2",
+        "@oclif/plugin-help": "^2.1.6",
+        "@oclif/plugin-not-found": "^1.2.2",
+        "@oclif/plugin-plugins": "^1.7.8",
+        "better-sqlite3": "^5.4.0",
+        "bip39": "^3.0.2",
+        "chalk": "^3.0.0",
+        "clear": "^0.1.0",
+        "cli-table3": "^0.5.1",
+        "cli-ux": "^5.2.1",
+        "dayjs": "^1.8.14",
+        "env-paths": "^2.2.0",
+        "execa": "^1.0.0",
+        "fast-levenshtein": "^2.0.6",
+        "fs-extra": "^8.0.1",
+        "got": "^9.6.0",
+        "is-reachable": "^4.0.0",
+        "json-buffer": "^3.0.1",
+        "latest-version": "^5.1.0",
+        "lodash.minby": "^4.6.0",
+        "lodash.sample": "^4.2.1",
+        "nodejs-tail": "^1.1.0",
+        "pino-pretty": "^3.1.0",
+        "pino": "^5.12.6",
+        "pretty-bytes": "^5.2.0",
+        "pretty-ms": "^5.0.0",
+        "prompts": "^2.1.0",
+        "read-last-lines": "^1.7.1",
+        "semver": "^7.0.0",
+        "sql": "^0.78.0",
+        "uuid": "^3.3.2",
+        "wif": "^2.0.6"
     },
     "devDependencies": {
+        "@oclif/dev-cli": "^1.22.0",
         "@sindresorhus/tsconfig": "^0.7.0",
+        "@types/execa": "^0.9.0",
+        "@types/fast-levenshtein": "^0.0.1",
+        "@types/fs-extra": "^8.0.0",
+        "@types/hapi__boom": "^7.4.0",
+        "@types/hapi__hapi": "^18.2.1",
         "@types/jest": "^24.0.15",
+        "@types/lodash.minby": "^4.6.6",
+        "@types/lodash.sample": "^4.2.6",
         "@types/node": "^12.6.2",
+        "@types/pino": "^5.20.0",
         "@types/prettier": "^1.16.4",
+        "@types/pretty-bytes": "^5.1.0",
+        "@types/pretty-ms": "^4.0.0",
+        "@types/prompts": "^2.4.0",
         "@types/rimraf": "^2.0.2",
+        "@types/semver": "^6.0.0",
+        "@types/uuid": "^3.4.4",
+        "@types/wif": "^2.0.1",
         "codecov": "^3.5.0",
         "cross-env": "^6.0.0",
         "del-cli": "^3.0.0",
         "husky": "^4.0.0",
-        "jest": "^24.8.0",
         "jest-extended": "^0.11.2",
+        "jest": "^24.8.0",
         "lerna": "^3.15.0",
         "lint-staged": "^9.2.0",
+        "nock": "^11.7.0",
         "prettier": "^1.18.2",
         "rimraf": "^3.0.0",
         "ts-jest": "^24.0.2",
-        "tslint": "^5.18.0",
         "tslint-config-prettier": "^1.18.0",
+        "tslint": "^5.18.0",
         "typescript": "^3.5.3"
     },
-    "workspaces": [
-        "packages/*"
-    ],
-    "husky": {
+    "engines": {
+        "node": ">=12.x"
+    },
+    "publishConfig": {
+        "access": "public"
+    },
+    "oclif": {
+        "commands": "./dist/commands",
         "hooks": {
-            "pre-commit": "lint-staged"
-        }
+            "init": [
+                "./dist/hooks/init/update"
+            ],
+            "command_not_found": [
+                "./dist/hooks/command_not_found/suggest"
+            ]
+        },
+        "bin": "exchange-json-rpc",
+        "plugins": [
+            "@oclif/plugin-autocomplete",
+            "@oclif/plugin-commands",
+            "@oclif/plugin-help"
+        ]
     }
 }
