--- packages/cli/package.json
@@ -1,131 +1,0 @@
-{
-    "name": "@arkecosystem/exchange-json-rpc-cli",
-    "version": "1.0.6",
-    "description": "A JSON-RPC 2.0 specification compliant server for Exchanges to interact with the ARK Blockchain.",
-    "license": "MIT",
-    "author": "Brian Faust <brian@ark.io>",
-    "files": [
-        "/bin",
-        "/dist",
-        "/oclif.manifest.json"
-    ],
-    "main": "dist/index",
-    "types": "dist/index",
-    "bin": {
-        "exchange-json-rpc": "./bin/run"
-    },
-    "scripts": {
-        "build": "yarn clean && yarn compile",
-        "build:watch": "yarn clean && yarn compile -w",
-        "clean": "del dist",
-        "compile": "../../node_modules/typescript/bin/tsc",
-        "exchange-json-rpc": "./bin/run",
-        "format": "yarn lint && yarn prettier",
-        "lint": "../../node_modules/tslint/bin/tslint -c ../../tslint.json -p ../../tslint-config.json './src/**/*.ts' --fix",
-        "prepack": "../../node_modules/.bin/oclif-dev manifest && npm shrinkwrap",
-        "postpack": "rm -f oclif.manifest.json",
-        "prepublishOnly": "yarn build",
-        "prettier": "prettier --write \"./*.{ts,js,json,md}\" \"./src/**/*.{ts,js,json,md}\" \"./__tests__/**/*.{ts,js,json,md}\""
-    },
-    "jest": {
-        "bail": true,
-        "collectCoverage": true,
-        "collectCoverageFrom": [
-            "src/**/*.ts",
-            "!src/index.ts",
-            "!**/node_modules/**"
-        ],
-        "coverageDirectory": "<rootDir>/.coverage",
-        "coverageReporters": [
-            "json",
-            "lcov",
-            "text",
-            "clover",
-            "html"
-        ],
-        "moduleFileExtensions": [
-            "ts",
-            "tsx",
-            "js",
-            "jsx",
-            "json",
-            "node"
-        ],
-        "setupFilesAfterEnv": [
-            "jest-extended"
-        ],
-        "testEnvironment": "node",
-        "testMatch": [
-            "**/*.test.ts"
-        ],
-        "transform": {
-            "^.+\\.tsx?$": "ts-jest"
-        },
-        "verbose": true,
-        "watchman": false
-    },
-    "dependencies": {
-        "@arkecosystem/crypto": "^2.4 || ^2.5",
-        "@arkecosystem/exchange-json-rpc": "^1.0.6",
-        "@arkecosystem/peers": "^0.2.1",
-        "@faustbrian/foreman": "^0.1.3",
-        "@oclif/command": "^1.5.14",
-        "@oclif/config": "^1.13.0",
-        "@oclif/plugin-autocomplete": "^0.1.0",
-        "@oclif/plugin-commands": "^1.2.2",
-        "@oclif/plugin-help": "^2.1.6",
-        "@oclif/plugin-not-found": "^1.2.2",
-        "@oclif/plugin-plugins": "^1.7.8",
-        "chalk": "^3.0.0",
-        "clear": "^0.1.0",
-        "cli-table3": "^0.5.1",
-        "cli-ux": "^5.2.1",
-        "dayjs": "^1.8.14",
-        "env-paths": "^2.2.0",
-        "execa": "^1.0.0",
-        "fast-levenshtein": "^2.0.6",
-        "fs-extra": "^8.0.1",
-        "latest-version": "^5.1.0",
-        "lodash.minby": "^4.6.0",
-        "nodejs-tail": "^1.1.0",
-        "pretty-bytes": "^5.2.0",
-        "pretty-ms": "^5.0.0",
-        "prompts": "^2.1.0",
-        "read-last-lines": "^1.7.1",
-        "semver": "^7.0.0"
-    },
-    "devDependencies": {
-        "@oclif/dev-cli": "^1.22.0",
-        "@types/execa": "^0.9.0",
-        "@types/fast-levenshtein": "^0.0.1",
-        "@types/fs-extra": "^8.0.0",
-        "@types/lodash.minby": "^4.6.6",
-        "@types/pretty-bytes": "^5.1.0",
-        "@types/pretty-ms": "^4.0.0",
-        "@types/prompts": "^2.4.0",
-        "@types/semver": "^6.0.0"
-    },
-    "engines": {
-        "node": ">=10.x"
-    },
-    "publishConfig": {
-        "access": "public"
-    },
-    "oclif": {
-        "commands": "./dist/commands",
-        "hooks": {
-            "init": [
-                "./dist/hooks/init/update"
-            ],
-            "command_not_found": [
-                "./dist/hooks/command_not_found/suggest"
-            ]
-        },
-        "bin": "exchange-json-rpc",
-        "plugins": [
-            "@oclif/plugin-autocomplete",
-            "@oclif/plugin-commands",
-            "@oclif/plugin-help"
-        ]
-    }
-}
