--- packages/rpc/package.json
@@ -1,58 +1,0 @@
-{
-    "name": "@arkecosystem/exchange-json-rpc",
-    "version": "1.0.6",
-    "description": "A JSON-RPC 2.0 specification compliant server for Exchanges to interact with the ARK Blockchain.",
-    "license": "MIT",
-    "author": "Brian Faust <brian@ark.io>",
-    "files": [
-        "/dist"
-    ],
-    "main": "dist/index",
-    "types": "dist/index",
-    "scripts": {
-        "build": "yarn clean && yarn compile",
-        "build:watch": "yarn clean && yarn compile -w",
-        "clean": "del dist",
-        "compile": "../../node_modules/typescript/bin/tsc",
-        "format": "yarn lint && yarn prettier",
-        "lint": "../../node_modules/tslint/bin/tslint -c ../../tslint.json -p ../../tslint-config.json './src/**/*.ts' --fix",
-        "prepublishOnly": "yarn build",
-        "prettier": "prettier --write \"./*.{ts,js,json,md}\" \"./src/**/*.{ts,js,json,md}\" \"./__tests__/**/*.{ts,js,json,md}\"",
-        "test": "jest"
-    },
-    "dependencies": {
-        "@arkecosystem/crypto": "^2.6.0-next.0",
-        "@hapi/boom": "^8.0.0",
-        "@hapi/hapi": "^18.3.1",
-        "@hapist/json-rpc": "^0.2.0",
-        "@hapist/whitelist": "^0.1.0",
-        "better-sqlite3": "^5.4.0",
-        "bip39": "^3.0.2",
-        "fs-extra": "^8.0.1",
-        "got": "^9.6.0",
-        "is-reachable": "^4.0.0",
-        "json-buffer": "^3.0.1",
-        "lodash.sample": "^4.2.1",
-        "pino": "^5.12.6",
-        "pino-pretty": "^3.1.0",
-        "sql": "^0.78.0",
-        "uuid": "^3.3.2",
-        "wif": "^2.0.6"
-    },
-    "devDependencies": {
-        "@types/fs-extra": "^8.0.0",
-        "@types/hapi__boom": "^7.4.0",
-        "@types/hapi__hapi": "^18.2.1",
-        "@types/lodash.sample": "^4.2.6",
-        "@types/pino": "^5.20.0",
-        "@types/uuid": "^3.4.4",
-        "@types/wif": "^2.0.1",
-        "nock": "^11.7.0"
-    },
-    "engines": {
-        "node": ">=10.x"
-    },
-    "publishConfig": {
-        "access": "public"
-    }
-}
