--- packages/rpc/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "outDir": "dist"
-    },
-    "include": ["src/**/**.ts"]
-}
