+++ src/commands/status.ts
