+++ src/commands/stop.ts
