+++ src/commands/update.ts
