+++ tsconfig.json
@@ -16,5 +16,6 @@
         "target": "es2018",
         "typeRoots": ["node_modules/@types"]
     },
+    "include": ["src/**/**.ts"],
     "exclude": ["node_modules", "**/*.spec.ts", "**/*.test.ts"]
 }
